requirements.txt
# Core dependencies
python-magic>=0.4.24
Pillow>=9.0.0
PyPDF2>=2.0.0
python-docx>=0.8.11
SpeechRecognition>=3.8.1
moviepy>=1.0.3
ffmpeg-python>=0.2.0
tqdm>=4.62.0
numpy>=1.21.0
pandas>=1.3.0

# Optional dependencies
ebooklib>=0.17.1
pydub>=0.25.1
imageio>=2.13.0
scikit-image>=0.19.0

# For parallel processing
concurrent futures backport for Python <3.8